fix: rename QuotaDeletedForTierIntegrationEvent to TierQuotaDefinitionDeletedIntegrationEvent
