+++ Modules/Quotas/src/Quotas.Application/IntegrationEvents/Outgoing/TierQuotaDefinitionDeletedIntegrationEvent.cs
@@ -1,9 +1,9 @@
 ﻿using Enmeshed.BuildingBlocks.Application.Abstractions.Infrastructure.EventBus.Events;
 
 namespace Backbone.Modules.Quotas.Application.IntegrationEvents.Outgoing;
-public class QuotaDeletedForTierIntegrationEvent : IntegrationEvent
+public class TierQuotaDefinitionDeletedIntegrationEvent : IntegrationEvent
 {
-    public QuotaDeletedForTierIntegrationEvent(string tierId, string tierQuotaDefinitionId) : base($"{tierQuotaDefinitionId}/Deleted")
+    public TierQuotaDefinitionDeletedIntegrationEvent(string tierId, string tierQuotaDefinitionId) : base($"{tierQuotaDefinitionId}/Deleted")
     {
         TierId = tierId;
         TierQuotaDefinitionId = tierQuotaDefinitionId;
