+++ Modules/Quotas/test/Quotas.Application.Tests/Tests/Identities/TierQuotaDefinitionDeletedIntegrationEventHandlerTests.cs
@@ -1,5 +1,5 @@
 ﻿using Backbone.Modules.Quotas.Application.Infrastructure.Persistence.Repository;
-using Backbone.Modules.Quotas.Application.IntegrationEvents.Incoming.QuotaDeletedForTier;
+using Backbone.Modules.Quotas.Application.IntegrationEvents.Incoming.TierQuotaDefinitionDeleted;
 using Backbone.Modules.Quotas.Application.IntegrationEvents.Outgoing;
 using Backbone.Modules.Quotas.Application.Tests.TestDoubles;
 using Backbone.Modules.Quotas.Domain.Aggregates.Identities;
@@ -10,7 +10,7 @@ using Microsoft.Extensions.Logging;
 using Xunit;
 
 namespace Backbone.Modules.Quotas.Application.Tests.Tests.Identities;
-public class QuotaDeletedForTierIntegrationEventHandlerTests
+public class TierQuotaDefinitionDeletedIntegrationEventHandlerTests
 {
     [Fact]
     public async void Deletes_tier_quota_after_consuming_integration_event()
@@ -35,7 +35,7 @@ public class QuotaDeletedForTierIntegrationEventHandlerTests
         var handler = CreateHandler(identitiesRepository, tierQuotaDefinitionsRepository);
 
         // Act
-        await handler.Handle(new QuotaDeletedForTierIntegrationEvent(tier.Id, tierQuotaDefinition.Id));
+        await handler.Handle(new TierQuotaDefinitionDeletedIntegrationEvent(tier.Id, tierQuotaDefinition.Id));
 
         // Assert
         A.CallTo(() => identitiesRepository.Update(A<IEnumerable<Identity>>.That.Matches(ids =>
@@ -44,9 +44,9 @@ public class QuotaDeletedForTierIntegrationEventHandlerTests
         ).MustHaveHappened();
     }
 
-    private QuotaDeletedForTierIntegrationEventHandler CreateHandler(IIdentitiesRepository identities, FindTierQuotaDefinitionsStubRepository tierQuotaDefinitions)
+    private TierQuotaDefinitionDeletedIntegrationEventHandler CreateHandler(IIdentitiesRepository identities, FindTierQuotaDefinitionsStubRepository tierQuotaDefinitions)
     {
-        var logger = A.Fake<ILogger<QuotaDeletedForTierIntegrationEventHandler>>();
-        return new QuotaDeletedForTierIntegrationEventHandler(identities, tierQuotaDefinitions, logger);
+        var logger = A.Fake<ILogger<TierQuotaDefinitionDeletedIntegrationEventHandler>>();
+        return new TierQuotaDefinitionDeletedIntegrationEventHandler(identities, tierQuotaDefinitions, logger);
     }
 }
